Revert merging of page and page content files
