+++ src/new-client/src/app/home-page.tsx
@@ -1,0 +1,10 @@
+import SiteList from '@components/site-list';
+import { Site } from '@types';
+
+interface HomePageProps {
+  sites: Site[];
+}
+
+export const HomePage = ({ sites }: HomePageProps) => {
+  return <SiteList sites={sites} />;
+};
