+++ src/new-client/src/app/page.tsx
@@ -1,8 +1,7 @@
 import { fetchUserProfile } from '@services/appointmentsService';
 import NhsPage from '@components/nhs-page';
 import { Metadata } from 'next';
-import SiteList from '@components/site-list';
-import { Site } from '@types';
+import { HomePage } from './home-page';
 
 export const metadata: Metadata = {
   title: 'Appointment Management Service',
@@ -17,14 +16,6 @@ const Page = async () => {
       <HomePage sites={userProfile?.availableSites ?? []} />
     </NhsPage>
   );
-};
-
-interface HomePageProps {
-  sites: Site[];
-}
-
-export const HomePage = ({ sites }: HomePageProps) => {
-  return <SiteList sites={sites} />;
 };
 
 export default Page;
