+++ src/new-client/src/app/site/[site]/users/manage/page.tsx
@@ -1,8 +1,5 @@
 ﻿import NhsPage from '@components/nhs-page';
-import NhsPageTitle from '@components/nhs-page-title';
-import { When } from '@components/when';
-import AssignRoles from './assign-roles';
-import FindUserForm from './find-user-form';
+import { ManageUsersPage } from './manage-users-page';
 
 export type UserPageProps = {
   params: {
@@ -31,32 +28,6 @@ const AssignRolesPage = ({ params, searchParams }: UserPageProps) => {
         searchParams={searchParams}
       />
     </NhsPage>
-  );
-};
-
-export const ManageUsersPage = ({
-  params,
-  searchParams,
-  userIsSpecified,
-}: UserPageProps & { userIsSpecified: boolean }) => {
-  return (
-    <div className="nhsuk-grid-row">
-      <div className="nhsuk-grid-column-one-half">
-        <div className="nhsuk-form-group">
-          <NhsPageTitle title="Staff Role Management" />
-          <div className="nhsuk-hint">
-            Set the details and roles of a new user
-          </div>
-        </div>
-
-        <When condition={userIsSpecified}>
-          <AssignRoles params={params} searchParams={searchParams} />
-        </When>
-        <When condition={!userIsSpecified}>
-          <FindUserForm site={params.site} />
-        </When>
-      </div>
-    </div>
   );
 };
