+++ src/new-client/src/app/site/[site]/users/users-page.test.tsx
@@ -1,5 +1,5 @@
 ﻿import { render, screen, waitFor } from '@testing-library/react';
-import { UsersPage } from './page';
+import { UsersPage } from './users-page';
 import { getMockUserAssignments, mockRoles } from '../../../testing/data';
 
 const mockSiteId = 'TEST';
